requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
tqdm>=4.62.0
tensorboard>=2.7.0
wandb>=0.12.0
albumentations>=1.1.0
efficientnet-pytorch>=0.7.0
timm>=0.6.0
pytest>=6.2.0
pytest-cov>=3.0.0
black>=21.0.0
flake8>=4.0.0
pre-commit>=2.15.0
